rework run command to be used as a library
